src/requirements.txt
pandas
numpy
scikit-learn
nltk
matplotlib
seaborn
tqdm
joblib
pytorch

